server/requirements.txt
django
djangorestframework
markdown 
django-filter
djoser
djangorestframework-simplejwt
django-cors-headers